Move redirect logic into logic file
