+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/content_sources/components/display_settings/display_settings.test.tsx
@@ -7,7 +7,6 @@
 
 import '../../../../../__mocks__/shallow_useeffect.mock';
 
-import { mockKibanaValues } from '../../../../../__mocks__';
 import { setMockValues, setMockActions } from '../../../../../__mocks__';
 import { exampleResult } from '../../../../__mocks__/content_sources.mock';
 
@@ -25,11 +24,11 @@ import { DisplaySettings } from './display_settings';
 import { FieldEditorModal } from './field_editor_modal';
 
 describe('DisplaySettings', () => {
-  const { navigateToUrl } = mockKibanaValues;
   const { exampleDocuments, searchResultConfig } = exampleResult;
   const initializeDisplaySettings = jest.fn();
   const setServerData = jest.fn();
   const setColorField = jest.fn();
+  const handleSelectedTabChanged = jest.fn();
 
   const values = {
     isOrganization: true,
@@ -46,6 +45,7 @@ describe('DisplaySettings', () => {
       initializeDisplaySettings,
       setServerData,
       setColorField,
+      handleSelectedTabChanged,
     });
     setMockValues({ ...values });
   });
@@ -83,7 +83,7 @@ describe('DisplaySettings', () => {
       const tabsEl = wrapper.find(EuiTabbedContent);
       tabsEl.prop('onTabClick')!(tabs[0]);
 
-      expect(navigateToUrl).toHaveBeenCalledWith('/sources/123/display_settings/');
+      expect(handleSelectedTabChanged).toHaveBeenCalledWith('search_results');
     });
 
     it('handles second tab click', () => {
@@ -91,7 +91,7 @@ describe('DisplaySettings', () => {
       const tabsEl = wrapper.find(EuiTabbedContent);
       tabsEl.prop('onTabClick')!(tabs[1]);
 
-      expect(navigateToUrl).toHaveBeenCalledWith('/sources/123/display_settings/result_detail');
+      expect(handleSelectedTabChanged).toHaveBeenCalledWith('result_detail');
     });
   });
