+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/content_sources/components/display_settings/display_settings.tsx
@@ -20,18 +20,10 @@ import {
 } from '@elastic/eui';
 
 import { clearFlashMessages } from '../../../../../shared/flash_messages';
-import { KibanaLogic } from '../../../../../shared/kibana';
 import { Loading } from '../../../../../shared/loading';
 import { UnsavedChangesPrompt } from '../../../../../shared/unsaved_changes_prompt';
-import { AppLogic } from '../../../../app_logic';
 import { ViewContentHeader } from '../../../../components/shared/view_content_header';
 import { SAVE_BUTTON } from '../../../../constants';
-
-import {
-  DISPLAY_SETTINGS_RESULT_DETAIL_PATH,
-  DISPLAY_SETTINGS_SEARCH_RESULT_PATH,
-  getContentSourcePath,
-} from '../../../../routes';
 
 import {
   UNSAVED_MESSAGE,
@@ -42,7 +34,7 @@ import {
   SEARCH_RESULTS_LABEL,
   RESULT_DETAIL_LABEL,
 } from './constants';
-import { DisplaySettingsLogic } from './display_settings_logic';
+import { DisplaySettingsLogic, TabId } from './display_settings_logic';
 import { FieldEditorModal } from './field_editor_modal';
 import { ResultDetail } from './result_detail';
 import { SearchResults } from './search_results';
@@ -52,17 +44,16 @@ interface DisplaySettingsProps {
 }
 
 export const DisplaySettings: React.FC<DisplaySettingsProps> = ({ tabId }) => {
-  const { initializeDisplaySettings, setServerData } = useActions(DisplaySettingsLogic);
+  const { initializeDisplaySettings, setServerData, handleSelectedTabChanged } = useActions(
+    DisplaySettingsLogic
+  );
 
   const {
     dataLoading,
-    sourceId,
     addFieldModalVisible,
     unsavedChanges,
     exampleDocuments,
   } = useValues(DisplaySettingsLogic);
-
-  const { isOrganization } = useValues(AppLogic);
 
   const hasDocuments = exampleDocuments.length > 0;
 
@@ -87,12 +78,7 @@ export const DisplaySettings: React.FC<DisplaySettingsProps> = ({ tabId }) => {
   ] as EuiTabbedContentTab[];
 
   const onSelectedTabChanged = (tab: EuiTabbedContentTab) => {
-    const path =
-      tab.id === tabs[1].id
-        ? getContentSourcePath(DISPLAY_SETTINGS_RESULT_DETAIL_PATH, sourceId, isOrganization)
-        : getContentSourcePath(DISPLAY_SETTINGS_SEARCH_RESULT_PATH, sourceId, isOrganization);
-
-    KibanaLogic.values.navigateToUrl(path);
+    handleSelectedTabChanged(tab.id as TabId);
   };
 
   const handleFormSubmit = (e: FormEvent) => {
